requirements.txt
# requirements.txt
streamlit
pillow
google-generativeai